refactor: move `hasAccessToHash` query out of auth info context
